catalog-service/src/main/java/com/adnan/catalog/dto/ExternalProductDto.java
package com.adnan.catalog.dto;

public class ExternalProductDto {
    
    private Number id;
    private String title;
    private String description;
    private String image;
    private Number price;

    public Number getId() {
        return id;
    }

    public void setId(Number id) {
        this.id = id;
    }

    public String getTitle() {
        return title;
    }

    public void setTitle(String title) {
        this.title = title;
    }

    public String getDescription() {
        return description;
    }

    public void setDescription(String description) {
        this.description = description;
    }

    public String getImage() {
        return image;
    }

    public void setImage(String image) {
        this.image = image;
    }

    public Number getPrice() {
        return price;
    }

    public void setPrice(Number price) {
        this.price = price;
    }
    
}
